Trials

+++ src/main/java/com/capgemini/IMCcalculator/entities/BMIEntity.java
@@ -2,6 +2,10 @@ package com.capgemini.IMCcalculator.entities;
 
 import lombok.*;
 
+import javax.validation.Valid;
+import javax.validation.constraints.DecimalMin;
+import javax.validation.constraints.Pattern;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
@@ -9,8 +13,20 @@ import lombok.*;
 @Setter
 public class BMIEntity {
 
-    private double height;
-    private double mass;
+//    @Pattern(
+//            regexp = "^[1-9]",
+//            message = "You should have entered numbers")
+//    @DecimalMin(value = "0",
+//            message = "Height should be greater than zero")
+    private int height;
+
+
+//    @Pattern(
+//            regexp = "^[1-9]",
+//            message = "You should have entered numbers")
+//    @DecimalMin(value = "0",
+//            message = "Mass should be greater than zero")
+    private int mass;
     private String message;
 
 }
